Trivial pyproject.toml change to test release job
This change is on a branch internal to my fork, to test the change
I am considering proposing to the if-condition on the job in the
release workflow. The file pyproject.toml must be modified by a PR
for the workflow to be triggered at all (both before and after my
change to the if condition), so something like this is necessary
for testing.